Bug removed with unspeakable levels of uglyness.
A bug somewhere in the map/resourcemanager caused the draw() call
in WindowManager to instead call prePhysicsStep() on objects
loaded from a map (but not others). I have no idea what caused
this, but tearing the needed code from Map and ResourceManger
and placing it in TestState worked. Assuming the vtable on the
affected objects got corrupted somehow.